+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/AnalyticsEventParams.kt
@@ -3,7 +3,7 @@ package com.braintreepayments.api.core
 import androidx.annotation.RestrictTo
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-data class AnalyticsEventParams(
+data class AnalyticsEventParams @JvmOverloads constructor(
     var payPalContextId: String? = null,
     var linkType: String? = null,
     var isVaultRequest: Boolean = false,